rename: 내 상품 컴포넌트명 수정
